Cleanup ENUM interface and add allow_null option
