+++ app/core/interfaces/enum/input.html
@@ -1,0 +1,14 @@
+<style>
+	.select2-results__option--highlighted[aria-selected] {
+		background-color: #eee;
+	}
+</style>
+
+<div class="select-container">
+	<select name="{{name}}" {{#if readonly}}disabled{{/if}}>
+		<option value="" {{#unless allowNull}}disabled{{/unless}}>{{t "select_from_below"}}</option>
+		{{#options}}
+		<option value="{{value}}" {{#if selected}}selected{{/if}}>{{value}}</option>
+		{{/options}}
+	</select>
+</div>
